break out attachment CSS to it's own module
