+++ app/javascript/mastodon/features/compose/redesign/poll.tsx
@@ -21,8 +21,8 @@ import {
 import { useAppSelector, useAppDispatch } from '@/mastodon/store';
 import { DAY, HOUR, MINUTE } from '@/mastodon/utils/time';
 
+import classes from './attachments.module.scss';
 import { selectComposePoll } from './selectors';
-import classes from './styles.module.scss';
 
 const messages = defineMessages({
   option_placeholder: {
