+++ app/javascript/mastodon/features/compose/redesign/styles.module.scss
@@ -232,137 +232,6 @@
   }
 }
 
-// Attachments
-
-.mediaSingle .mediaUpload {
-  min-width: 120px;
-  max-width: 100%;
-  max-height: var(--max-media-height-large);
-}
-
-.mediaGrid {
-  display: grid;
-  gap: var(--space-3xs);
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: repeat(2, 1fr);
-  aspect-ratio: 3/2;
-
-  &[data-number='2'] {
-    grid-template-rows: 1fr;
-  }
-
-  &[data-number='3'] .mediaUpload:first-child {
-    grid-row: span 2;
-  }
-}
-
-.mediaUpload {
-  border-radius: var(--radius-md);
-  position: relative;
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-
-.mediaUploadPending {
-  animation: pulse 1s ease-in-out infinite;
-  background-color: var(--color-bg-disabled);
-}
-
-@keyframes pulse {
-  0% {
-    opacity: 1;
-  }
-
-  50% {
-    opacity: 0.8;
-  }
-}
-
-.mediaMenuButton {
-  position: absolute;
-  top: var(--space-xs);
-  right: var(--space-xs);
-}
-
-.mediaAlt {
-  @include mixins.type-micro;
-
-  display: block;
-  padding: var(--space-2xs);
-  border-radius: var(--radius-xs);
-  background: var(--color-bg-inverted);
-  color: var(--color-text-inverted);
-  position: absolute;
-  bottom: var(--space-xs);
-  right: var(--space-xs);
-  text-transform: uppercase;
-}
-
-.mediaMenuDelete {
-  color: var(--color-text-error);
-}
-
-// Polls
-
-.poll {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-sm);
-  border: 0.5px solid var(--color-text-tertiary);
-  border-radius: var(--radius-md);
-  padding: var(--space-xs);
-
-  ol {
-    display: flex;
-    flex-direction: column;
-    gap: var(--space-xs);
-    counter-reset: poll;
-    width: 100%;
-  }
-
-  label {
-    @include mixins.type-label-sm;
-  }
-}
-
-.pollOption {
-  display: flex;
-  gap: var(--space-sm);
-  align-items: center;
-  counter-increment: poll;
-
-  &::before {
-    @include mixins.type-micro;
-
-    content: counter(poll, upper-alpha);
-    width: var(--space-lg);
-    height: var(--space-lg);
-    flex-shrink: 0;
-    line-height: var(--space-lg);
-    text-align: center;
-    border-radius: var(--radius-xs);
-    background: var(--color-bg-highlight);
-  }
-}
-
-.pollAddNew {
-  margin-left: var(--space-lg);
-}
-
-.pollMultipleToggle {
-  align-self: flex-start;
-}
-
-.pollControls {
-  display: flex;
-  justify-content: space-between;
-}
-
-.pollDurationSelect {
-  text-align: center;
-}
-
 // Emoji picker
 
 div.emojiRoot {
